requirements.txt
pymongo
python-dotenv
langchain-core
langchain-openai
langchain-mongodb
langchain_voyageai
langchain_community
pypdf